requirements/prod.txt
# Everything needed in production

# Flask
Flask==1.1.2
Werkzeug==1.0.1
click>=7.0

# Database
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.23
psycopg2-binary==2.8.6

# Migrations
Flask-Migrate==2.7.0

# Forms
email-validator==1.1.2
Flask-WTF==0.14.3
WTForms==2.3.3

# Deployment
gevent==21.1.2
gunicorn>=19.9.0
supervisor==4.2.2

# Flask Static Digest
Flask-Static-Digest==0.2.1

# Auth
Flask-Login==0.5.0
Flask-Bcrypt==0.7.1

# Caching
Flask-Caching>=1.7.2

# Images processing
numpy
opencv-python
pillow
scipy
scikit-image

# Debug toolbar
Flask-DebugToolbar==0.11.0

# Environment variable parsing
environs==9.3.1
